requirements.txt
python-dateutil
setuptools
wheel
pyindi-client
picamera